requirements.txt
accelerate==1.4.0
aiohttp==3.11.13
av==14.2.0
bitsandbytes==0.41.0
datasets==2.16.1
decord==0.6.0
deepspeed==0.14.4
einops==0.6.1
einops-exts==0.0.4
ffmpeg-python==0.2.0
flash-attn==2.6.3
fsspec==2023.10.0
ftfy==6.3.1
gdown==5.2.0
GitPython==3.1.44
gradio_client==0.2.9
huggingface-hub==0.29.1
imageio==2.37.0
joblib==1.4.2
markdown2==2.5.3
numpy==1.26.1
opencv-python==4.11.0.86
pandas==2.2.3
peft==0.4.0
pillow==10.2.0
protobuf==3.19.6
psutil==7.0.0
pyarrow==19.0.1
python-dateutil==2.9.0.post0
PyYAML==6.0.1
regex==2024.11.6
requests==2.31.0
scikit-learn==1.2.2
scipy==1.15.2
sentencepiece==0.1.99
tokenizers==0.21.1
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tqdm==4.67.1
transformers==4.49.0
triton==2.1.0
tyro==0.9.16
uvicorn==0.34.0
wandb==0.18.7
websockets==15.0
xxhash==3.5.0
